feat: :sparkles: Add and Refactor serializers
Add some new serializers classes to the project and organize the directory structure of these serializers for greater organization
