+++ app/projac/views.py
@@ -2,9 +2,15 @@
 
 from rest_framework import viewsets
 from projac.models import Projeto
-from projac.serializers import ProjetoSerializer
+from .serializers import ProjetoListSerializer, ProjetoDetailSerializer
 
 class ProjetoViewSet(viewsets.ModelViewSet):
     """Projeto viewset"""
     queryset = Projeto.objects.all()
-    serializer_class = ProjetoSerializer
+
+    def get_serializer_class(self):
+        """Get serializer class method"""
+        if self.action == "list":
+            return ProjetoListSerializer
+        return ProjetoDetailSerializer
+
